fix(refactor): move geometry into it's own directory / component
